Revert "Merged in ALFREDAPI-480_remote (pull request #18)"
We want to use the stored commit hash that is committed in the
repository, and not always automatically update to the latest version
that is available on the configured submodule branch.

Projects with a single release branch have multiple references to that
same release branch, but with different checked out commits for
different versions (probably on the tag of that version).
We don't want all those references to be automatically updated to the
latest version, but we want to keep them pointing to the existing commit
that they point at.

This reverts commit ee9f2c6b9fa5f056ea1bb4322fde629702a26864, reversing
changes made to 74ad39820c5ec7b7648667d77ee2dd8790955ac5.